+++ app/client/doc/CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.234] - 2023-03-201
+## [0.0.0.24] - 2023-03-23
 
-- Updated documentation on CLI modes
+- Make `PromptPrintNodeState` the first prompt in debug mode
+- Minor cleanup to documentation related to CLI modes
 - Fixed one logging statement
 
 ## [0.0.0.23] - 2023-03-21