+++ p2p/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.39] - 2023-03-23
+
+- Variable name and comment improvements
+
 ## [0.0.0.35] - 2023-03-21
 
 - Add log for `StateMachineTransitionEvent`
@@ -96,6 +100,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.0.19] - 2023-01-19
 
 - Rewrite `interface{}` to `any`
+
 ## [0.0.0.18] - 2023-01-11
 
 - Add a lock to the mempool to avoid parallel messages which has caused the node to crash in the past