+++ shared/CHANGELOG.md
@@ -7,13 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.43] - 2023-03-23
+
+- Add a deadline to the primary event handling to get visibility into concurrency issues
+
 ## [0.0.0.42] - 2023-03-21
 
 - Add `TransitionEventToMap()` helper function for logging
 
 ## [0.0.0.41] - 2023-03-21
 
-- Added *temporary* `shared/p2p` package to hold P2P interfaces common to both legacy and libp2p modules
+- Added _temporary_ `shared/p2p` package to hold P2P interfaces common to both legacy and libp2p modules
 - Added `Peerstore` interface
 - Added `Peer` and `PeerList` and interfaces
 - Moved `typesP2P.AddrBookMap` to `sharedP2P.PeerAddrMap` and refactor to implement the new `Peerstore` interface